requirements.txt
appdirs==1.4.4
black==21.7b0
click==8.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
regex==2021.8.21
tomli==1.2.1
